Update HoaDonModel constructor, add TrangThai to SELECT statement in BanHangService
The commit includes changes to the HoaDonModel class, where a new constructor with the parameter "trangThai" is added. In the BanHangService class, the SELECT statement in the "getAllHD2" method is updated to include the "TrangThai" column in the SELECT clause and GROUP BY clause. Additionally, the new constructor of HoaDonModel is used in the result set mapping in the "getAllHD2" method.