Update FormBanHang.form and FormBanHang.java to swap positions of "Tất cả" and "Chờ thanh toán" in ComboBoxModel
Change the order of elements in the ComboBoxModel used in FormBanHang.form and FormBanHang.java to swap the positions of "Tất cả" and "Chờ thanh toán". Also, update the corresponding Vietnamese strings in FormBanHang.java.

Add PdfPTable for product details in exportToPDF method of FormBanHang.java

Add a new PdfPTable to display product details (tên, giá, số lượng, thành tiền) in the exportToPDF method of FormBanHang.java. Create a separate function addProductTableHeader to add table headers and another function addProductTableRow to add product details to the table.

Update BanHangService.java to include getHoaDonChoThanhToan method

Add a new method getHoaDonChoThanhToan to BanHangService.java to retrieve a list of HoaDonModel objects with the status "Chờ thanh toán". This method was previously implemented in the constructor of FormBanHang but has been moved to the service layer for better separation of concerns.